@@ -77,87 +80,83 @@
     "react-transition-group": "^1.1.1",
     "redux": "^3.x.x",
     "reselect": "^2.5.4",
-    "swagger-client": "^3.9.0",
-    "swagger-ui": "^3.22.3",
+    "swagger-client": "^3.9.5",
+    "swagger-ui": "^3.24.0",
     "traverse": "^0.6.6",
     "whatwg-fetch": "^2.0.3",
     "yaml-js": "^0.1.4"
   },
   "devDependencies": {
-    "autoprefixer": "6.6.1",
-    "babel-core": "^6.23.1",
-    "babel-eslint": "^6.1.2",
-    "babel-loader": "^6.3.2",
+    "@babel/cli": "^7.5.5",
+    "@babel/core": "^7.0.0",
+    "@babel/plugin-proposal-class-properties": "^7.5.5",
+    "@babel/plugin-proposal-optional-chaining": "^7.2.0",
+    "@babel/plugin-transform-runtime": "^7.5.5",
+    "@babel/preset-env": "^7.0.0",
+    "@babel/preset-react": "^7.0.0",
+    "@babel/register": "^7.5.5",
+    "@babel/runtime-corejs2": "^7.0.0",
+    "@release-it/conventional-changelog": "^1.1.0",
+    "autoprefixer": "^6.6.1",
+    "babel-eslint": "^10.0.2",
+    "babel-loader": "^8.0.6",
     "babel-plugin-module-alias": "^1.6.0",
-    "babel-plugin-transform-async-to-generator": "^6.24.1",
+    "babel-plugin-module-resolver": "^3.2.0",
     "babel-plugin-transform-es2015-constants": "^6.1.4",
-    "babel-plugin-transform-runtime": "^6.23.0",
-    "babel-preset-es2015": "^6.22.0",
-    "babel-preset-es2015-ie": "^6.6.2",
-    "babel-preset-react": "^6.23.0",
-    "babel-preset-stage-0": "^6.22.0",
-    "babel-runtime": "^6.23.0",
+    "babel-plugin-transform-react-remove-prop-types": "^0.4.24",
     "bundlesize": "^0.17.2",
     "cross-env": "^5.1.3",
-    "css-loader": "0.22.0",
-    "cypress": "^3.1.5",
-    "deep-extend": "^0.4.1",
+    "css-loader": "^3.1.0",
+    "cssnano": "^4.1.10",
+    "cypress": "^3.4.1",
+    "deep-extend": "^0.6.0",
     "deepmerge": "^1.3.2",
     "enzyme": "^3.3.0",
     "enzyme-adapter-react-15": "^1.0.5",
-    "eslint": "^2.13.1",
+    "eslint": "^6.1.0",
     "eslint-plugin-import": "^2.13.0",
     "eslint-plugin-react": "^4.3.0",
     "expect": "^1.20.2",
-    "expose-loader": "^0.7.3",
     "extract-text-webpack-plugin": "0.8.2",
     "file-loader": "0.8.4",
     "git-describe": "^4.0.1",
     "html-webpack-plugin": "^2.28.0",
-    "http-server": "^0.9.0",
-    "imports-loader": "0.6.5",
+    "http-server": "^0.11.1",
+    "ignore-assets-webpack-plugin": "^2.0.1",
     "jsdom": "11.12.0",
     "jsdom-global": "3.0.2",
     "json-loader": "0.5.3",
-    "karma": "^0.13.22",
-    "karma-chrome-launcher": "^0.2.3",
-    "karma-mocha": "^0.2.2",
-    "karma-sourcemap-loader": "^0.3.7",
-    "karma-webpack": "1.8.0",
-    "less": "2.5.3",
-    "less-loader": "2.2.1",
+    "less": "^3.9.0",
+    "less-loader": "^5.0.0",
     "license-checker": "^8.0.4",
     "matcher": "^0.1.2",
-    "mocha": "^3.2.0",
+    "mini-css-extract-plugin": "^0.8.0",
+    "mocha": "^6.2.0",
     "mocha-jsdom": "^2.0.0",
+    "npm-audit-ci-wrapper": "^2.3.0",
     "npm-run-all": "^4.1.3",
     "null-loader": "0.1.1",
-    "open": "0.0.5",
-    "postcss-loader": "0.7.0",
+    "open": "^6.4.0",
+    "postcss-loader": "^3.0.0",
     "raw-loader": "0.5.1",
     "react-hot-loader": "^1.3.1",
     "react-test-renderer": "^15.6.2",
-    "release-it": "^7.6.1",
+    "release-it": "^12.3.5",
     "rewiremock": "^3.7.4",
     "rimraf": "^2.6.0",
     "standard": "^8.6.0",
-    "standard-loader": "^5.0.0",
     "style-loader": "0.13.0",
     "uglifyjs-webpack-plugin": "^1.2.5",
-    "url-loader": "0.5.6",
-    "webpack": "^1.15.0",
+    "url-loader": "^2.1.0",
+    "webpack": "^4.36.1",
     "webpack-bundle-size-analyzer": "^2.5.0",
-    "webpack-dev-server": "1.14.0",
-    "worker-loader": "^0.7.1"
+    "webpack-cli": "^3.3.6",
+    "webpack-dev-server": "^3.7.2",
+    "worker-loader": "^2.0.0"
   },
   "config": {
     "deps_check_dir": ".deps_check"
   },
-  "browserslist": [
-    "> 1%",
-    "last 2 versions",
-    "IE 10"
-  ],
   "bundlesize": [
     {
       "path": "./dist/swagger-editor-bundle.js",
