@@ -1,7 +1,6 @@
 import deepMerge from "deepmerge"
 import SwaggerUI from "swagger-ui"
 import EditorLayout from "./layout"
-import "swagger-ui/dist/swagger-ui.css"
 
 import EditorPlugin from "./plugins/editor"
 import LocalStoragePlugin from "./plugins/local-storage"
@@ -41,7 +40,8 @@ const plugins = {
 }
 
 const defaults = {
-  dom_id: "#swagger-editor", // eslint-disable-line camelcase, we have this prop for legacy reasons.
+  // we have the `dom_id` prop for legacy reasons
+  dom_id: "#swagger-editor", // eslint-disable-line camelcase
   layout: "EditorLayout",
   presets: [
     SwaggerUI.presets.apis
@@ -56,7 +56,7 @@ const defaults = {
   swagger2ConverterUrl: "https://converter.swagger.io/api/convert",
 }
 
-module.exports = function SwaggerEditor(options) {
+export default function SwaggerEditor(options) {
   let mergedOptions = deepMerge(defaults, options)
 
   mergedOptions.presets = defaults.presets.concat(options.presets || [])
@@ -64,4 +64,4 @@ module.exports = function SwaggerEditor(options) {
   return SwaggerUI(mergedOptions)
 }
 
-module.exports.plugins = plugins
+SwaggerEditor.plugins = plugins
